drafted stream for reading chunks
This will be used to read chunks from the local socket and forward them to the remote end. We will need a similar thing that can take chunks from the remote end and push them to the local socket. For that case, I'll probably just make a Future that writes a single chunk (buffering happens on the OS layer, anyway).